pyproject.toml
[project]
name = "AI-TRADING"
version = "0.1.0"
description = "AI-powered Bitcoin trading system"
authors = [
    { name = "khaimai" }
]
dependencies = [
    "pyyaml",
    "pandas",
    "numpy",
    "scikit-learn",
    "ccxt",
    "matplotlib",
    "seaborn",
    "pyarrow"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
